+++ packages/generator-teams/src/app/GeneratorTeamsAppOptions.ts
@@ -14,7 +14,6 @@ export class GeneratorTeamsAppOptions {
     name: string;
     shouldUseSubDir: boolean;
     libraryName: string;
-    'skip-install': string;
     telemetry: boolean;
     namespace: string;
     developer: string;
@@ -34,6 +33,7 @@ export class GeneratorTeamsAppOptions {
     updateManifestVersion: boolean = false; // Set to true when asked to update manifest
     useAzureAppInsights: boolean = false;
     azureAppInsightsKey?: string = undefined;
+    lintingSupport: boolean = true;
 
     /* Tabs */
     tab: boolean;