+++ packages/generator-teams/src/app/templates/webpack.config.js
@@ -3,9 +3,8 @@
 // Licensed under the MIT license.
 
 const webpack = require("webpack");
-const Dotenv = require("dotenv-webpack");
 const nodeExternals = require("webpack-node-externals");
-const ESLintPlugin = require("eslint-webpack-plugin");
+<% if (lintingSupport) { %>const ESLintPlugin = require("eslint-webpack-plugin");<% } %>
 
 const path = require("path");
 const fs = require("fs");
@@ -75,16 +74,14 @@ const config = [{
         }]
     },
     plugins: [
-        new Dotenv({
-            systemvars: true
-        })
+        new webpack.EnvironmentPlugin({ PUBLIC_HOSTNAME: undefined, TAB_APP_ID: null, TAB_APP_URI: null })
     ]
 }
 ];
-
+<% if (lintingSupport) { %>
 if (lint !== false) {
     config[0].plugins.push(new ESLintPlugin({ extensions: ["ts", "tsx"], failOnError: false }));
     config[1].plugins.push(new ESLintPlugin({ extensions: ["ts", "tsx"], failOnError: false }));
 }
-
+<% } %>
 module.exports = config;