+++ packages/generator-teams/src/connector/templates/src/server/{connectorName}/{connectorComponentName}.ts
@@ -75,7 +75,7 @@ export class <%=connectorComponentName%> implements IConnector {
                         {
                             activityTitle: "Ping",
                             activityText: "Sample ping ",
-                            activityImage: `https://${process.env.HOSTNAME}/assets/icon.png`,
+                            activityImage: `https://${process.env.PUBLIC_HOSTNAME}/assets/icon.png`,
                             facts: [
                                 {
                                     name: "Generator",