+++ packages/generator-teams/tests/MessageExtensionGenerator.spec.ts
@@ -9,22 +9,25 @@ const tests = require("./tests.json");
 
 describe('teams:messageExtension', function () {
 
-    const MESSAGEEXTSION_HTML_FILES = [
+    const MESSAGE_EXTENSION_HTML_FILES = [
         'src/public/teamsSolutionMessageExtension/config.html'
     ];
 
-    const MESSAGEEXTSION_SCRIPT_FILES = [
+    const MESSAGE_EXTENSION_SCRIPT_FILES = [
         'src/client/teamsSolutionMessageExtension/TeamsSolutionMessageExtensionConfig.tsx'
     ];
 
-    const MESSAGEEXTSION_SCRIPT_TEST_FILES = [
+    const MESSAGE_EXTENSION_SCRIPT_TEST_FILES = [
         'src/client/teamsSolutionMessageExtension/__tests__/TeamsSolutionMessageExtensionConfig.spec.tsx'
     ];
 
-    const MESSAGEEXTSION_FILES = [
+    const MESSAGE_EXTENSION_SERVER_TEST_FILES = [
+        'src/server/teamsSolutionMessageExtension/__tests__/TeamsSolutionMessageExtension.spec.ts'
+    ];
+    const MESSAGE_EXTENSION_FILES = [
         'src/server/teamsSolutionMessageExtension/TeamsSolutionMessageExtension.ts'
     ];
-    const MESSAGEEXTSION__TASKMODULE_FILES = [
+    const MESSAGE_EXTENSION_TASK_MODULE_FILES = [
         'src/client/mteamsSolutionMessageExtension/TeamsSolutionMessageExtensionAction.tsx',
         'src/public/teamsSolutionMessageExtension/action.html'
     ];
@@ -41,10 +44,10 @@ describe('teams:messageExtension', function () {
                 })
             });
             it("Should have messaging extension html files", async () => {
-                assert.noFile(MESSAGEEXTSION_HTML_FILES);
+                assert.noFile(MESSAGE_EXTENSION_HTML_FILES);
             });
             it("Should have messaging extension client files", async () => {
-                assert.noFile(MESSAGEEXTSION_SCRIPT_FILES);
+                assert.noFile(MESSAGE_EXTENSION_SCRIPT_FILES);
             });
         } else {
             it("Should have a Compose Extension with update", async () => {
@@ -53,24 +56,34 @@ describe('teams:messageExtension', function () {
                 })
             });
             it("Should have messaging extension html files", async () => {
-                assert.file(MESSAGEEXTSION_HTML_FILES);
+                assert.file(MESSAGE_EXTENSION_HTML_FILES);
             });
             it("Should have messaging extension client files", async () => {
-                assert.file(MESSAGEEXTSION_SCRIPT_FILES);
+                assert.file(MESSAGE_EXTENSION_SCRIPT_FILES);
             });
         }
 
         it("Should have messaging extension app files", async () => {
-            assert.file(MESSAGEEXTSION_FILES);
+            assert.file(MESSAGE_EXTENSION_FILES);
         });
 
         if (prompts.unitTestsEnabled && prompts.messagingExtensionCanUpdateConfiguration != false) {
             it("Should have messaging extension unit test files", async () => {
-                assert.file(MESSAGEEXTSION_SCRIPT_TEST_FILES);
+                assert.file(MESSAGE_EXTENSION_SCRIPT_TEST_FILES);
             });
         } else {
             it("Should not have messaging extension unit test files", async () => {
-                assert.noFile(MESSAGEEXTSION_SCRIPT_TEST_FILES);
+                assert.noFile(MESSAGE_EXTENSION_SCRIPT_TEST_FILES);
+            });
+        }
+
+        if (prompts.unitTestsEnabled) {
+            it("Should have middleware messaging extension unit test files", async () => {
+                assert.file(MESSAGE_EXTENSION_SERVER_TEST_FILES);
+            });
+        } else {
+            it("Should not have middleware messaging extension unit test files", async () => {
+                assert.noFile(MESSAGE_EXTENSION_SERVER_TEST_FILES);
             });
         }
 
@@ -79,26 +92,26 @@ describe('teams:messageExtension', function () {
                 assert.jsonFileContent('src/manifest/manifest.json', { composeExtensions: [{ commands: [{ type: "action" }] }] });
             });
             it("Middleware should define onSubmitAction", async () => {
-                assert.fileContent(MESSAGEEXTSION_FILES[0], "public async onSubmitAction(");
+                assert.fileContent(MESSAGE_EXTENSION_FILES[0], "public async onSubmitAction(");
             });
             it("Manifest should not declare a query messaging extension", async () => {
                 assert.noJsonFileContent('src/manifest/manifest.json',{ composeExtensions: [{ commands: [{ type: "query" }] }] });
             });
             it("Bot middleware should not have the onQuery method", async () => {
-                assert.noFileContent(MESSAGEEXTSION_FILES[0], "public async onQuery(");
+                assert.noFileContent(MESSAGE_EXTENSION_FILES[0], "public async onQuery(");
             });
         } else if (prompts.messagingExtensionType == "query") {
             it("Manifest should declare a query messaging extension", async () => {
                 assert.jsonFileContent('src/manifest/manifest.json',{ composeExtensions: [{ commands: [{ type: "query" }] }] });
             });
             it("Bot middleware should have the onQuery method", async () => {
-                assert.fileContent(MESSAGEEXTSION_FILES[0], "public async onQuery(");
+                assert.fileContent(MESSAGE_EXTENSION_FILES[0], "public async onQuery(");
             });
             it("Manifest should declare an action messaging extension", async () => {
                 assert.noJsonFileContent('src/manifest/manifest.json', { composeExtensions: [{ commands: [{ type: "action" }] }] });
             });
             it("Middleware should define onSubmitAction", async () => {
-                assert.noFileContent(MESSAGEEXTSION_FILES[0], "public async onSubmitAction(");
+                assert.noFileContent(MESSAGE_EXTENSION_FILES[0], "public async onSubmitAction(");
             });
         } else {
             it("Manifest should have empty commands", async () => {
@@ -108,7 +121,7 @@ describe('teams:messageExtension', function () {
                 assert.jsonFileContent('src/manifest/manifest.json', { composeExtensions: [{ messageHandlers: [{ type: "link" }] }] });
             });
             it("Bot middleware should have the onQueryLink method", async () => {
-                assert.fileContent(MESSAGEEXTSION_FILES[0], "public async onQueryLink(");
+                assert.fileContent(MESSAGE_EXTENSION_FILES[0], "public async onQueryLink(");
             });
         }
         if (prompts.messagingExtensionType == "action" && prompts.messagingExtensionActionInputType == "static") {
@@ -116,7 +129,7 @@ describe('teams:messageExtension', function () {
                 assert.jsonFileContent('src/manifest/manifest.json', { composeExtensions: [{ commands: [{ fetchTask: false }] }] });
             });
             it("Middleware should not define onFetchTask", async () => {
-                assert.noFileContent(MESSAGEEXTSION_FILES[0], "public async onFetchTask(");
+                assert.noFileContent(MESSAGE_EXTENSION_FILES[0], "public async onFetchTask(");
             });
         }
         if (prompts.messagingExtensionType == "action" && prompts.messagingExtensionActionInputType == "adaptiveCard") {
@@ -124,7 +137,7 @@ describe('teams:messageExtension', function () {
                 assert.jsonFileContent('src/manifest/manifest.json', { composeExtensions: [{ commands: [{ fetchTask: true }] }] });
             });
             it("Middleware should not define onFetchTask", async () => {
-                assert.fileContent(MESSAGEEXTSION_FILES[0], "public async onFetchTask(");
+                assert.fileContent(MESSAGE_EXTENSION_FILES[0], "public async onFetchTask(");
             });
         }
         if (prompts.messagingExtensionType != "queryLink") {