+++ packages/yoteams-build-core/src/index.ts
@@ -5,6 +5,8 @@
 import GulpClient from "gulp";
 import Undertaker from "undertaker";
 import { buildTasks } from "./buildTasks";
+import { codespacesTasks } from "./codespacesTasks";
+import { loadPlugins } from "./loadPlugins";
 import { manifest } from "./manifest";
 import { ngrokTasks } from "./ngrokTasks";
 import { nukeTasks } from "./nukeTasks";
@@ -45,4 +47,8 @@ export const setup = (gulp: GulpClient.Gulp, config: any): void => {
     buildTasks(gulp, config);
     serveTasks(gulp, config);
     ngrokTasks(gulp, config);
+    codespacesTasks(gulp, config);
+
+    // load additional tasks
+    loadPlugins(gulp, config);
 };