+++ packages/yoteams-build-core/src/ngrokTasks.ts
@@ -21,13 +21,13 @@ export const ngrokTasks = (gulp: GulpClient.Gulp, config: any) => {
         ngrok.connect(conf).then((url) => {
             log("[NGROK] Url: " + url);
             if (!conf.authtoken) {
-                log("[NGROK] You have been assigned a random ngrok URL that will only be available for this session. You wil need to re-upload the Teams manifest next time you run this command.");
+                log("[NGROK] You have been assigned a random ngrok URL that will only be available for this session. You will need to re-upload the Teams manifest next time you run this command.");
             }
             let hostName = url.replace("http://", "");
             hostName = hostName.replace("https://", "");
 
-            log("[NGROK] HOSTNAME: " + hostName);
-            process.env.HOSTNAME = hostName;
+            log("[NGROK] Public url: " + hostName);
+            process.env.PUBLIC_HOSTNAME = hostName;
 
             cb();