+++ packages/yoteams-build-core/CHANGELOG.md
@@ -5,11 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [*Unreleased*] 
+## [*1.2.0*]- <*2021-06-07*>
+
+### Added
+
+* Added `codespaces-serve` task, only available inside a Github Codespaces container, for automatic host name generation when using Github Codespaces
+* Watches for changes in `.env` files and rebuilds and restarts the solution, as well as scaffolds an updated manifest package
+* Support for plugins
 
 ### Changes
 
+* [BREAKING] Changed the environment variable `HOSTNAME` to `PUBLIC_HOSTNAME` to not reuse the system set HOSTNAME.
 * Updated referenced packages
+* Switched to `gulp-dart-sass` due to deprecated `node-sass`
+* Node 10.x no longer supported
+
+### Fixes
+
+* Fixed typo (#222)
 
 ## [*1.1.0*] - <*2021-04-15*>